requirements.txt
absl-py==0.11.0
astunparse==1.6.3
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.2.1
cssselect==1.1.0
dataclasses==0.8
filelock==3.0.12
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-images-download==2.8.0
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
hyperlink==20.0.1
idna==2.10
ImageScraper==2.0.7
importlib-metadata==2.0.0
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
joblib==0.17.0
Keras-Preprocessing==1.1.2
lxml==4.6.1
Markdown==3.3.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
parsel==1.6.0
pdftotext==2.1.5
Pillow==8.0.1
pkg-resources==0.0.0
Protego==0.1.16
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-docx==0.8.10
queuelib==1.5.0
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scipy==1.4.1
Scrapy==2.4.0
selenium==3.141.0
sentencepiece==0.1.91
service-identity==18.1.0
setproctitle==1.1.10
SimplePool==0.1
six==1.15.0
soupsieve==2.0.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tokenizers==0.9.3
tqdm==4.51.0
transformers==3.5.1
Twisted==20.3.0
urllib3==1.26.2
w3lib==1.22.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
